requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-rest-auth==7.0.1
Django==4.2.19
django-allauth==65.4.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==1.7.1
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
